Reference

FITS Tables Extensions

Base.names — Method

names(hdu::FitsTableHDU)
hdu.column_names

yield a vector of the column names of the FITS table extension in hdu.

Base.read — Method

read(io::IO, T)

Read a single value of type T from io, in canonical binary representation.

Note that Julia does not convert the endianness for you. Use ntoh or ltoh for this purpose.

read(io::IO, String)

Read the entirety of io, as a String (see also readchomp).

Examples

julia> io = IOBuffer("JuliaLang is a GitHub organization");

julia> read(io, Char)
'J': ASCII/Unicode U+004A (category Lu: Letter, uppercase)

julia> io = IOBuffer("JuliaLang is a GitHub organization");

julia> read(io, String)
"JuliaLang is a GitHub organization"

read([R=Array,] hdu::FitsTableHDU, col[, rows]; kwds...) -> vals::R

reads column col of the FITS table extension in hdu and returns its values and, possibly, its units.

The column col may be specified by its name or by its number. If col is a string or a symbol, keyword case specifies whether the case of letters matters (case = false by default).

Optional leading argument R is to specify the type of the result which can be an array type, to only retrieve the column values, or a tuple of array and string types, to retrieve the column values and their units.

Optional argument rows is to specify which rows to read. It can be an integer to read a single row, a unit range of integers to read these rows, or a colon : to read all rows (the default). Use hdu.first_row and hdu.last_row to retrieve the first and last row numbers.

See read!(::DenseArray,::FitsTableHDU,::ColumName) for the other possible keywords.

Base.read — Method

read([Dict,] hdu::FitsTableHDU[, cols[, rows]]) -> dict

reads some columns of the FITS table extension in hdu as a dictionary indexed by the column names. The columns to read can be specified by cols which may be a single column name/index, a tuple/range/vector of column names/numbers, or a colon : to read all columns (the default). Column names may be strings or symbols (not a mixture of these). The rows to read can be specified by rows as a single row index, a unit range of row numbers, or a colon : to read all rows (the default).

Keyword rename is to specify a function to change column names. If unspecified, the colmun names are left unchanged if keyword case is true and converted to uppercase letters otherwise.

Keyword units can be used to indicate whether to retrieve the units of the columns. If units is String, the values of the dictionary will be 2-tuples (data,units) with data the column data and units the column units as a string. Otherwise, if units is nothing (the default), the values of the dictionary will just be the columns data.

To avoid the units keyword, the following methods are provided:

read(Dict{String,Array},               hdu[, cols[, rows]])
read(Dict{String,Tuple{Array,String}}, hdu[, cols[, rows]])

to yield the same result as read(hdu,...) with respectively units=nothing and units=String.

Base.read — Method

read(Vector, hdu::FitsTableHDU[, cols[, rows]]) -> vec::Vector

reads some columns of the FITS table extension in hdu as a vector. The columns to read can be specified by cols which may be a single column name/index, a tuple/range/vector of column names/numbers, or a colon : to read all columns (the default). Column names may be strings or symbols (not a mixture of these). The rows to read can be specified by rows as a single row index, a unit range of row numbers, or a colon : to read all rows (the default). V is the type of the result.

Keyword units can be used to indicate whether to retrieve the units of the columns. If units is String, the values of the dictionary will be 2-tuples (data,units) with data the column data and units the column units as a string. Otherwise, if units=nothing (the default), the values of the dictionary will just be the columns data.

To avoid the units keyword and allow more control on the type of the result, the following 2 methods are provided:

read(Vector{<:Array},               hdu[, cols[, rows]])
read(Vector{Tuple{<:Array,String}}, hdu[, cols[, rows]])

to yield the same result as read(hdu,...) with respectively units=nothing and units=String.

Base.read! — Method

read!(arr::DenseArray, hdu::FitsTableHDU, col) -> arr

overwrites the elements of array arr with values of the column col of the FITS table extension in hdu and returns arr.

The column col may be specified by its name or by its number. If col is a string or a symbol, keyword case indicates whether the case of letters matters (case = false by default).

Keyword first may be specified with the index of the first row to read. By default, first = hdu.first_row and reading starts at the first row of the table.

Keyword anynull may be specified with a reference to a boolean (Ref{Bool}()) to retrieve whether any of the read values is undefined.

Keyword null may be specified with a reference to a value of the same type as the elements of the destination arr (Ref{eltype(arr)}()) to retrieve the value of undefined values. Keyword null may also be set with an array of Bool of same size as arr and which will be set to true for undefined values and to false elsewhere.

Output arrays arr and null must have contiguous elements, in other words, they must be dense arrays.

Base.read! — Method

read!(dict, hdu::FitsTableHDU[, cols[, rows]]) -> dict

overwrites the contents of the dictionary dict with the colum(s) cols read from the FITS table extension in hdu and returns the dictionary. Any previous contents is erased, call merge!(dict,hdu,...) to preserve contents.

Base.merge! — Method

merge!(dict, hdu::FitsTableHDU[, cols[, rows]]) -> dict

merges the contents of the dictionary dict with the colum(s) cols read from the FITS table extension in hdu and returns the dictionary.

Base.push! — Method

push!(vec, hdu::FitsTableHDU[, cols[, rows]]) -> vec

appends rows rows of columns cols read from FITS table extension hdu to the vector vec and returns it.

FITS Image Extensions

Base.read — Method

read(R::Type = Array, hdu::FitsImageHDU) -> arr::R

reads the FITS image extension in hdu. Optional argument R is to restrict the ouput type and improve type-stability. For example:

arr = convert(Array{Float32}, read(hdu))
arr = read(Array{Float32}, hdu)
arr = read(Array{Float32,2}, hdu)

yield similar results if hdu is a 2-D image extension but the 2nd example is more efficient than the 1st one as no temporary array is needed if the pixel type is not equivalent to Float32 and the 3rd example is completely type-stable.

Keywords anynull and null may be specified to deal with undefined pixel values (see documentation for read!(::DenseArray{<:Number},::FitsImageHDU,::SubArrayIndex...)).

read(R::Type = Array, hdu::FitsImageHDU, inds...) -> arr::R

reads a rectangular sub-region of the FITS image extension in hdu defined by the indices inds.... Keywords anynull and null may be specified to deal with undefined pixel values (see documentation for read!(::DenseArray{<:Number},::FitsImageHDU,::SubArrayIndex...)).

The result is similar to:

arr = read(R, hdu)[inds...]

but is generally more efficient as no array other than the result is allocated and fewer values are read.

Base.read! — Method

read!(arr::DenseArray{<:Number}, hdu::FitsImageHDU, inds...) -> arr

overwrites arr with a rectangular sub-region of the FITS image extension in hdu defined by the indices inds.... The destination array must have the same dimensions as the rectangular sub-region to read (the same rules as for sub-indexing an array are applied to determine the dimensions of the sub-region). Keywords anynull and null may be specified to deal with undefined pixel values.

Missing docstring.

Missing docstring for read!(::DenseArray, ::FitsImageHDU). Check Documenter's build log for details.

Base.write — Method

write(file::FitsFile, FitsImageHDU{T}, dims...=()) -> hdu
write(file::FitsFile, FitsImageHDU, T::Type=UInt8, dims...=()) -> hdu
write(file::FitsFile, FitsImageHDU, bitpix::Integer, dims=()) -> hdu

create a new primary array or image extension in FITS file file with a specified pixel type T and size dims.... If the FITS file is currently empty then a primary array is created, otherwise a new image extension is appended to the file. Pixel type can be specified as a numeric type T or as an integer BITPIX code bitpix.

An object to manage the new extension is returned which can be used to push header cards and then to write the data.

For example:

hdu = write(file, FitsImageHDU, eltype(arr), size(arr))
hdu["KEY1"] = val1             # add a 1st header record
hdu["KEY2"] = (val2, str2)     # add a 2nd header record
hdu["KEY3"] = (nothing, str3)  # add a 3rd header record
write(hdu, arr)                # write data

will create a new Header Data Unit (HDU) storing array arr with 3 additional header records: one named "KEY1" with value val1 and no comments, another named "KEY2" with value val2 and comment string str2, and yet another one named "KEY3" with no value and with comment string str3. Note that special names "COMMENT", "HISTORY", and "" indicating commentary entries have no associated, only a comment string, say str which can be specified as str or as (,str).

Base.write — Method

write(file::FitsFile, hdr, arr::AbstractArray) -> file

writes a new FITS Image Extension in file with non-structural header keywords specified by hdr and data specified by array arr.

Base.write — Method

write(hdu::FitsImageHDU, arr::AbstractArray{<:Number}) -> hdu

writes all the elements of the array arr to the pixels of the FITS image extension of the header data unit hdu. The element of arr are converted to the type of the pixels. The default is to write the complete image pixels, so the size of the destination arr must be identical to the dimensions of the FITS image extension. This behavior may be changed by specifying another value than nothing for the keywords first and/or last:

To write a rectangular sub-image, specify keywords first and last with the coordinates of the first and last pixels to write as an N-tuple of integers, with N the number of dimensions of the FITS image extension.

To write consecutive pixels, specify at least one of the keywords first and/or last with the index of the first and/or last pixels to write as an integer.

When at least one of the keywords first and/or last is not nothing, the dimensions of arr are not considered. In any case, the number of elements of arr must be equal to the number of pixels to write.

Unless writing a rectangular sub-image, keyword null may be used to specify the value of undefined elements in arr. For integer FITS images, the FITS null value is defined by the BLANK keyword (an error is returned if the BLANK keyword doesn't exist). For floating point FITS images the special IEEE NaN (Not-a-Number) value will be written into the FITS file.